Refactor large files under 800-line checkstyle limit (#125)
- Extract OCI protocol methods from RepoManager (1039→711) into OciRegistryClient
- Extract number parsing from Parser (918→761) into NumberParser
- Split CollectionFunctions (870→27) into ListFunctions + DictFunctions
- Lower checkstyle FileLength max from 1000 to 800
- Remove FileLength suppressions for RepoManager and Parser

Closes #125
