+++ checkstyle-suppressions.xml
@@ -31,9 +31,8 @@
     <!-- TrustManager interface implementations must be public; ignore visibility rule -->
     <suppress checks="SpringMethodVisibility" files="KpsComparisonTest\.java"/>
 
-    <!-- RepoManager has deeply nested YAML parsing logic and many methods that push it over size limits -->
+    <!-- RepoManager has deeply nested YAML parsing logic -->
     <suppress checks="NestedIfDepth" files="RepoManager\.java"/>
-    <suppress checks="FileLength" files="RepoManager\.java"/>
     <suppress checks="MethodLength" files="RepoManager\.java"/>
 
     <!-- Engine.renderWithSubcharts is slightly over limit due to log guards -->
