+++ jhelm-core/src/test/java/org/alexmond/jhelm/core/service/RepoManagerTest.java
@@ -17,8 +17,6 @@ import org.junit.jupiter.api.io.TempDir;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.mockito.stubbing.Answer;
-import tools.jackson.databind.JsonNode;
-import tools.jackson.databind.json.JsonMapper;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
@@ -174,50 +172,6 @@ class RepoManagerTest {
 		repoManager.verifyBlobDigest(file, "md5:abc123");
 	}
 
-	@Test
-	void testIsManifestIndexDetectsIndex() throws Exception {
-		JsonMapper mapper = JsonMapper.builder().build();
-		RepoManager repoManager = new RepoManager();
-
-		JsonNode index = mapper.readTree("""
-				{"mediaType":"application/vnd.oci.image.index.v1+json","manifests":[{"digest":"sha256:abc"}]}
-				""");
-		assertTrue(repoManager.isManifestIndex(index));
-
-		JsonNode singleManifest = mapper.readTree("""
-				{"mediaType":"application/vnd.oci.image.manifest.v1+json","layers":[]}
-				""");
-		assertFalse(repoManager.isManifestIndex(singleManifest));
-	}
-
-	@Test
-	void testResolveDigestFromIndexPrefersNoPlatform() throws Exception {
-		JsonMapper mapper = JsonMapper.builder().build();
-		RepoManager repoManager = new RepoManager();
-
-		JsonNode index = mapper.readTree("""
-				{"manifests":[
-				  {"digest":"sha256:amd64","platform":{"os":"linux","architecture":"amd64"}},
-				  {"digest":"sha256:noplat"}
-				]}
-				""");
-		assertEquals("sha256:noplat", repoManager.resolveDigestFromIndex(index));
-	}
-
-	@Test
-	void testResolveDigestFromIndexPrefersLinuxAmd64() throws Exception {
-		JsonMapper mapper = JsonMapper.builder().build();
-		RepoManager repoManager = new RepoManager();
-
-		JsonNode index = mapper.readTree("""
-				{"manifests":[
-				  {"digest":"sha256:arm64","platform":{"os":"linux","architecture":"arm64"}},
-				  {"digest":"sha256:amd64","platform":{"os":"linux","architecture":"amd64"}}
-				]}
-				""");
-		assertEquals("sha256:amd64", repoManager.resolveDigestFromIndex(index));
-	}
-
 	// ── Config file operations (use temp configPath constructor) ─────────────
 
 	@Test
@@ -280,30 +234,6 @@ class RepoManagerTest {
 	}
 
 	// ── Package-private logic methods ────────────────────────────────────────
-
-	@Test
-	void testParseOciUrlWithTag() throws IOException {
-		RepoManager rm = new RepoManager();
-		String[] parts = rm.parseOciUrl("oci://my.registry.io/charts/mychart:1.2.3");
-		assertEquals("my.registry.io", parts[0]);
-		assertEquals("charts/mychart", parts[1]);
-		assertEquals("1.2.3", parts[2]);
-	}
-
-	@Test
-	void testParseOciUrlDefaultsToLatest() throws IOException {
-		RepoManager rm = new RepoManager();
-		String[] parts = rm.parseOciUrl("oci://my.registry.io/charts/mychart");
-		assertEquals("my.registry.io", parts[0]);
-		assertEquals("charts/mychart", parts[1]);
-		assertEquals("latest", parts[2]);
-	}
-
-	@Test
-	void testParseOciUrlInvalid() {
-		RepoManager rm = new RepoManager();
-		assertThrows(IOException.class, () -> rm.parseOciUrl("oci://nopath"));
-	}
 
 	@ParameterizedTest
 	@CsvSource({ "'', 'https://charts.example.com/mychart-1.0.0.tgz'",
